6.4.2
Updated API:
-- unauthorizeClientLibrary was changed from an escaping signature with a completion block to a direct return signature.
This is due to the fact that it's calling a synchronous edgeEngine shutdown call internally and in fact is of no asynchronous nature.

-- edgeEngineServiceLink
-- edgeEngineWebSocketServiceLink
Changed both signatures to return a String, instead of a optional String. Since it is in fact guaranteed that a String can and will be returned.

Renamed APIs:
-- unauthorizeClientLibrary was renamed to unauthorizeClientLibrarySynchronously with a Swift redirect suggestion
-- platformServiceLink was renamed to edgeEngineServiceLink with a Swift redirect suggestion
-- platformWorkingDirectory was renamed to edgeEngineWorkingDirectory  with a Swift redirect suggestion
-- platformWebSocketServiceLink was renamed to edgeEngineWebSocketServiceLink with a Swift redirect suggestion

Updated API documentation
--  startEdgeEngine
--  stopEdgeEngineSynchronously